refactor: 完善多包边界与发布验收
