+++ .changeset/add-milkdown-package.md
@@ -1,0 +1,5 @@
+---
+'@cherry-markdown/milkdown': minor
+---
+
+新增独立 Milkdown 编辑器适配包，并对无法保留的 Cherry 语法返回明确错误。
