+++ .changeset/add-stream-package.md
@@ -1,0 +1,5 @@
+---
+'@cherry-markdown/stream': minor
+---
+
+新增仅依赖 Engine 与 Preview 的流式 Markdown 渲染包。
