+++ .changeset/legacy-package-forwarding.md
@@ -1,0 +1,5 @@
+---
+'cherry-markdown': minor
+---
+
+根编辑器改为组合新包，并在 0.x 保留可于 1.0 统一删除的历史产物转发入口。
