+++ .changeset/new-multipackages.md
@@ -1,9 +1,5 @@
 ---
-'@cherry-markdown/engine': minor
 '@cherry-markdown/preview': minor
-'@cherry-markdown/stream': minor
-'@cherry-markdown/milkdown': minor
-'cherry-markdown': minor
 ---
 
-按职责建立 engine / preview / stream / milkdown 独立包；根编辑器改为组合新包，并在 0.x 保留可于 1.0 统一删除的历史产物转发入口。
+新增独立 Preview 组合包，负责 Markdown 驱动的预览 DOM 生命周期。
