--- packages/engine/src/index.with-mermaid.js
@@ -1,30 +1,0 @@
-/**
- * Copyright (C) 2021 Tencent.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import CherryEngine from './index';
-import MermaidCodeEngine from './addons/cherry-code-block-mermaid-plugin';
-import mermaid from 'mermaid';
-
-// v9 及以下有 mermaid.mermaidAPI，v10+ 为 undefined
-const mermaidAPI = mermaid?.mermaidAPI;
-CherryEngine.usePlugin(MermaidCodeEngine, {
-  mermaid,
-  mermaidAPI,
-  theme: 'default',
-  sequence: { useMaxWidth: false, showSequenceNumbers: true },
-});
-
-export * from './index';
-export default CherryEngine;
