+++ packages/engine/test/architecture.spec.ts
@@ -1,5 +1,7 @@
 import { beforeEach, describe, expect, it } from 'vite-plus/test';
 import CherryEngine, { createSyntaxHook, Sanitizer } from '../src/index';
+import { existsSync, readFileSync } from 'node:fs';
+import { resolve } from 'node:path';
 
 describe('@cherry-markdown/engine architecture contract', () => {
   beforeEach(() => {
@@ -22,5 +24,13 @@ describe('@cherry-markdown/engine architecture contract', () => {
 
     expect(clean).toContain('<p>safe</p>');
     expect(clean).not.toContain('<script>');
+  });
+
+  it('keeps editor interaction and large renderer plugins outside Engine', () => {
+    const suggester = readFileSync(resolve(import.meta.dirname, '../src/syntax/hooks/Suggester.js'), 'utf8');
+    expect(suggester).not.toContain('@codemirror/');
+    expect(suggester).not.toContain('suggester-panel');
+    expect(suggester).not.toContain('onCodeMirrorChange');
+    expect(existsSync(resolve(import.meta.dirname, '../src/addons/cherry-code-block-mermaid-plugin.js'))).toBe(false);
   });
 });
