+++ packages/preview/types/env.d.ts
@@ -1,4 +1,5 @@
 declare const BUILD_ENV: string;
+declare module '*.css';
 interface Window {
   DOMParser?: typeof DOMParser;
 }
